feat: add multiple update* methods (#149)
* feat: add updateAuthenticators, updateVotingStrategies and updateSettings

* feat: add updateAuthenticatorsAndVotingStrategies

* fix: add --via-ir to forge coverage

* fix: add via_ir in foundry.toml

* refactor: split the functions into updateSettings and updateStrategies

* fix: check for metadatauriupdated event when testing

* fix: check for _min and _max and do not use internal setters

* refactor: move back from votingStrategiesMetadataURIs to votingStrategyMetadataURIs

* refactor: remove set* entrypoints

* refactor: remove EmptyArray check in internal functions

* fix: Differentiate NO_UPDATE_METADATA_URI and NO_UPDATE_METADATA_URI_HASH

* fix: fix NO_UPDATE_METADATA_URI_HASH

* fix: remove useless comments

* refactor: changing magic values for NO_UPDATE

* refactor: use types instead of variables for no_update

* refactor: use input struct for update settings

* fix: remove extra comments on UpdateSettingsInput

* chore: remove extra comment in updateSettings function

* chore: add comments on values of NO_UPDATE_*

* chore: fix typo in Space.col comments

* chore: updated comments

---------

+++ src/interfaces/space/ISpaceEvents.sol
@@ -116,7 +116,7 @@ interface ISpaceEvents {
 
     /// @notice Emitted when the voting delay is updated.
     /// @param newVotingDelay The new voting delay.
-    event VotingDelayUpdated(uint256 newVotingDelay);
+    event VotingDelayUpdated(uint32 newVotingDelay);
 
     /// @notice Emitted when a proposal is updated.
     /// @param proposalId The proposal id.
